backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
uvicorn==0.24.0.post1
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
boto3==1.34.0
Pillow==10.1.0
django-imagekit==5.0.2
django-unfold==0.28.0
djangorestframework-simplejwt==5.3.2
drf-spectacular==0.26.5
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ratelimit==4.1.0
sentry-sdk==1.39.1
cryptography==41.0.7
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
djangorestframework-jwt==1.11.0
